requirements.txt
pandas
numpy
geopy
requests
matplotlib
pandas-gbq
google-cloud-bigquery

